requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
PyQt5>=5.15.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
